Move find_name_in_file to document_symbol.rs
